These vars were in fact being set. It's not what we actually want.
These variables *were* being respected, causing the upstream Docker
images to think that the `jekyll` user was in fact the `root` user.
This is explicitly not desired behavior according to the upstream
authors and so is our bug. Fix it by not automatically setting these
variables in the environment on each GitHub Actions run.